BASELINE: Update Chromium to 69.0.3497.70

+++ chromium/components/sync_sessions/session_sync_bridge.cc
@@ -13,6 +13,8 @@
 #include "base/bind.h"
 #include "base/bind_helpers.h"
 #include "base/location.h"
+#include "base/sequenced_task_runner.h"
+#include "base/threading/sequenced_task_runner_handle.h"
 #include "base/time/time.h"
 #include "components/sync/base/hash_util.h"
 #include "components/sync/base/time.h"
@@ -94,20 +96,6 @@ class LocalSessionWriteBatch : public LocalSessionEventHandlerImpl::WriteBatch {
   syncer::ModelTypeChangeProcessor* const processor_;
 };
 
-bool IsSessionRestoreInProgress(SyncSessionsClient* sessions_client) {
-  DCHECK(sessions_client);
-  SyncedWindowDelegatesGetter* synced_window_getter =
-      sessions_client->GetSyncedWindowDelegatesGetter();
-  SyncedWindowDelegatesGetter::SyncedWindowDelegateMap windows =
-      synced_window_getter->GetSyncedWindowDelegates();
-  for (const auto& window_iter_pair : windows) {
-    if (window_iter_pair.second->IsSessionRestoreInProgress()) {
-      return true;
-    }
-  }
-  return false;
-}
-
 }  // namespace
 
 SessionSyncBridge::SessionSyncBridge(
@@ -132,8 +120,7 @@ SessionSyncBridge::SessionSyncBridge(
           store_factory,
           base::BindRepeating(&FaviconCache::UpdateMappingsFromForeignTab,
                               base::Unretained(&favicon_cache_)))),
-      is_session_restore_in_progress_(
-          IsSessionRestoreInProgress(sessions_client)) {
+      weak_ptr_factory_(this) {
   DCHECK(sessions_client_);
   DCHECK(local_session_event_router_);
   DCHECK(foreign_sessions_updated_callback_);
@@ -149,9 +136,9 @@ void SessionSyncBridge::ScheduleGarbageCollection() {
   if (!syncing_) {
     return;
   }
-  base::ThreadTaskRunnerHandle::Get()->PostTask(
+  base::SequencedTaskRunnerHandle::Get()->PostTask(
       FROM_HERE, base::BindOnce(&SessionSyncBridge::DoGarbageCollection,
-                                base::AsWeakPtr(this)));
+                                weak_ptr_factory_.GetWeakPtr()));
 }
 
 FaviconCache* SessionSyncBridge::GetFaviconCache() {
@@ -167,14 +154,6 @@ OpenTabsUIDelegate* SessionSyncBridge::GetOpenTabsUIDelegate() {
     return nullptr;
   }
   return syncing_->open_tabs_ui_delegate.get();
-}
-
-void SessionSyncBridge::OnSessionRestoreComplete() {
-  is_session_restore_in_progress_ = false;
-
-  if (syncing_) {
-    StartLocalSessionEventHandler();
-  }
 }
 
 syncer::SyncableService* SessionSyncBridge::GetSyncableService() {
@@ -196,9 +175,7 @@ base::Optional<syncer::ModelError> SessionSyncBridge::MergeSyncData(
   DCHECK(syncing_);
   DCHECK(change_processor()->IsTrackingMetadata());
 
-  if (!is_session_restore_in_progress_) {
-    StartLocalSessionEventHandler();
-  }
+  StartLocalSessionEventHandler();
 
   return ApplySyncChanges(std::move(metadata_change_list),
                           std::move(entity_data));
@@ -209,19 +186,11 @@ void SessionSyncBridge::StartLocalSessionEventHandler() {
   DCHECK(syncing_);
   DCHECK(!syncing_->local_session_event_handler);
   DCHECK(change_processor()->IsTrackingMetadata());
-  DCHECK(!is_session_restore_in_progress_);
 
-  // TODO(crbug.com/681921): Remove injecting |local_session_write_batch| and
-  // let the impl create one via the delegate once the directory-based
-  // implementation is removed.
-  std::unique_ptr<LocalSessionEventHandlerImpl::WriteBatch>
-      local_session_write_batch = CreateLocalSessionWriteBatch();
   syncing_->local_session_event_handler =
       std::make_unique<LocalSessionEventHandlerImpl>(
           /*delegate=*/this, sessions_client_,
-          syncing_->store->mutable_tracker(), local_session_write_batch.get());
-
-  local_session_write_batch->Commit();
+          syncing_->store->mutable_tracker());
 
   // Start processing local changes, which will be propagated to the store as
   // well as the processor.
@@ -295,9 +264,10 @@ base::Optional<syncer::ModelError> SessionSyncBridge::ApplySyncChanges(
       ->TransferChangesTo(batch->GetMetadataChangeList());
   SessionStore::WriteBatch::Commit(std::move(batch));
 
-  // This might overtrigger because we don't check if the batch is empty, but
-  // observers should handle these events well so we don't bother detecting.
-  foreign_sessions_updated_callback_.Run();
+  if (!entity_changes.empty()) {
+    foreign_sessions_updated_callback_.Run();
+  }
+
   return base::nullopt;
 }
 
@@ -307,7 +277,7 @@ void SessionSyncBridge::GetData(StorageKeyList storage_keys,
   std::move(callback).Run(syncing_->store->GetSessionDataForKeys(storage_keys));
 }
 
-void SessionSyncBridge::GetAllData(DataCallback callback) {
+void SessionSyncBridge::GetAllDataForDebugging(DataCallback callback) {
   DCHECK(syncing_);
   std::move(callback).Run(syncing_->store->GetAllSessionData());
 }
@@ -325,15 +295,14 @@ std::string SessionSyncBridge::GetStorageKey(
   return SessionStore::GetStorageKey(entity_data.specifics.session());
 }
 
-ModelTypeSyncBridge::DisableSyncResponse
-SessionSyncBridge::ApplyDisableSyncChanges(
+ModelTypeSyncBridge::StopSyncResponse SessionSyncBridge::ApplyStopSyncChanges(
     std::unique_ptr<MetadataChangeList> delete_metadata_change_list) {
   local_session_event_router_->Stop();
-  if (syncing_) {
+  if (syncing_ && delete_metadata_change_list) {
     syncing_->store->DeleteAllDataAndMetadata();
   }
   syncing_.reset();
-  return DisableSyncResponse::kModelNoLongerReadyToSync;
+  return StopSyncResponse::kModelNoLongerReadyToSync;
 }
 
 std::unique_ptr<LocalSessionEventHandlerImpl::WriteBatch>
@@ -347,9 +316,9 @@ SessionSyncBridge::CreateLocalSessionWriteBatch() {
     syncing_->local_data_out_of_sync = false;
     // We use PostTask() to avoid interferring with the ongoing handling of
     // local changes that triggered this function.
-    base::ThreadTaskRunnerHandle::Get()->PostTask(
+    base::SequencedTaskRunnerHandle::Get()->PostTask(
         FROM_HERE, base::BindOnce(&SessionSyncBridge::ResubmitLocalSession,
-                                  base::AsWeakPtr(this)));
+                                  weak_ptr_factory_.GetWeakPtr()));
   }
 
   return std::make_unique<LocalSessionWriteBatch>(
@@ -371,11 +340,12 @@ void SessionSyncBridge::OnFaviconVisited(const GURL& page_url,
   favicon_cache_.OnFaviconVisited(page_url, favicon_url);
 }
 
-void SessionSyncBridge::OnSyncStarting() {
+void SessionSyncBridge::OnSyncStarting(
+    const syncer::DataTypeActivationRequest& request) {
   DCHECK(!syncing_);
 
   session_store_factory_.Run(base::BindOnce(
-      &SessionSyncBridge::OnStoreInitialized, base::AsWeakPtr(this)));
+      &SessionSyncBridge::OnStoreInitialized, weak_ptr_factory_.GetWeakPtr()));
 }
 
 void SessionSyncBridge::OnStoreInitialized(
@@ -403,8 +373,7 @@ void SessionSyncBridge::OnStoreInitialized(
 
   // If initial sync was already done, MergeSyncData() will never be called so
   // we need to start syncing local changes.
-  if (change_processor()->IsTrackingMetadata() &&
-      !is_session_restore_in_progress_) {
+  if (change_processor()->IsTrackingMetadata()) {
     StartLocalSessionEventHandler();
   }
 }
@@ -481,8 +450,8 @@ std::unique_ptr<SessionStore::WriteBatch>
 SessionSyncBridge::CreateSessionStoreWriteBatch() {
   DCHECK(syncing_);
 
-  return syncing_->store->CreateWriteBatch(
-      base::BindOnce(&SessionSyncBridge::ReportError, base::AsWeakPtr(this)));
+  return syncing_->store->CreateWriteBatch(base::BindOnce(
+      &SessionSyncBridge::ReportError, weak_ptr_factory_.GetWeakPtr()));
 }
 
 void SessionSyncBridge::ResubmitLocalSession() {
